src/test/java/com/wkdwoo/book/springboot/web/HelloControllerTest.java
package com.wkdwoo.book.springboot.web;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@WebMvcTest(controllers = HelloController.class)
public class HelloControllerTest {

    @Autowired
    private MockMvc mvc;

    @Test
    public void hello가_리턴된다() throws Exception {
        String hello = "hello";

        mvc.perform(get("/hello"))
                .andExpect(status().isOk())
                .andExpect(content().string(hello));
    }

    /**
     * # param
     * - API 테스트할 때 사용될 요청 파라미터를 설정한다.
     * - 단, 값은 String만 허용된다.
     * - 그래서 숫자/ 날짜 등의 데이터도 등록할 때는 문자열로 변경해야만 가능하다.
     *
     * # jsonPath
     * - JSON 응닶값을 필드별로 검증할 수 있는 메소드
     * - $를 기준으로 필드명을 명시한다.
     * - 여기서는 name과 amount를 검증하니 $.name, $.amount로 검증한다.
     */
    @Test
    public void helloDto가_리턴된다() throws Exception {
        String name = "hello";
        int amount = 1000;


        mvc.perform(get("/hello/dto")
                .param("name", name)
                .param("amount", String.valueOf(amount)))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.name", is(name)))
                .andExpect(jsonPath("$.amount", is(amount)));

    }


}